Big code refactor
Using clang-tidy and clang-format
Removed unused stuff
Extracted imgui to class

+++ Game/Source/Entities/Ball/ballDefault.h
@@ -9,7 +9,7 @@ namespace Entities {
         inline static const float speed = 2.f;
         inline static const float scalarFixer = 0.006f;
         inline static const float timeout = 0.2f;
-        inline static const int   angle = 5.f;
+        inline static const float   angle = 5.f;
         inline static const glm::vec3 direction{ CAMERA.GetCameraFront() };
         inline static const glm::vec3 position{ 0.f, 0.f, 0.f };
     };
